Clean up
Removed "short" version. Will create a new repository for that.
It's easier to manage that way.

Removed 3dsvm label and FIM ideal files intended for use at the
scanner (within TABS). Planning to create a new branch for these files,
so they can be added as a git submodule within TABS (tabs/bin/labels)
and synced easily when they are updated.